// src/types/bindings/PdfTemplate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../../../src-tauri/bindings/serde_json/JsonValue";
import type { PdfDocType } from "./PdfDocType";
import type { PdfTemplateType } from "./PdfTemplateType";

export type PdfTemplate = { id: string, doc_type: PdfDocType, name: string, type: PdfTemplateType, file_path: string, config_json: JsonValue, created_at: string, updated_at: string, };